chore: Refactor P2000Api class and P2000Sensor class
Refactor the `P2000Api` class to rename the `filter` parameter to `apiFilter` in the `get_data` method. Also, update the `P2000Sensor` class to use the `apiFilter` parameter instead of the deprecated `filter` parameter. Additionally, add support for configuring the sensor icon in the `PLATFORM_SCHEMA` by introducing the `CONF_ICON` constant.